requirements.txt
affine==2.4.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-functions==1.18.0
cachetools==5.3.3
certifi==2024.2.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
color-operations==0.1.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
morecantile==5.3.0
numexpr==2.10.0
numpy==1.26.4
pillow==10.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
pyproj==3.6.1
pystac==1.10.0
python-dateutil==2.9.0.post0
rasterio==1.3.9
rio-tiler==6.4.4
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
typing_extensions==4.10.0
